+++ Core/TgStorage/Domain/Contacts/TgEfContactDto.cs
@@ -76,7 +76,26 @@ public sealed partial class TgEfContactDto : TgDtoBase, ITgDto<TgEfContactEntity
 
 	public override string ToString() => $"{Id} | {AccessHash}";
 
-	public TgEfContactDto Copy(TgEfContactDto dto, bool isUidCopy)
+    public override string ToConsoleString()
+    {
+        var name = TgDataFormatUtils.GetFormatString(FirstName, 30).TrimEnd();
+        name += TgDataFormatUtils.GetFormatString(LastName, 30).TrimEnd();
+        return 
+            $"{Id,11} | " +
+            $"{TgDataFormatUtils.GetFormatString(UserName, 25).TrimEnd(),-25} | " +
+            $"{(IsActive ? "active" : ""),-6} | " +
+            $"{TgDataFormatUtils.GetFormatString(PhoneNumber, 11).TrimEnd(),-11} | " +
+            $"{name,-40}";
+    }
+
+    public override string ToConsoleHeaderString() =>
+        $"{nameof(Id),11} | " +
+        $"{TgDataFormatUtils.GetFormatString(nameof(UserName), 25).TrimEnd(),-25} | " +
+        $"Active | " +
+        $"{TgDataFormatUtils.GetFormatString(nameof(PhoneNumber), 11).TrimEnd(),-11} | " +
+        $"Name";
+
+    public TgEfContactDto Copy(TgEfContactDto dto, bool isUidCopy)
 	{
 		base.Copy(dto, isUidCopy);
 		DtChanged = dto.DtChanged;