+++ Core/TgStorage/Domain/Filters/TgEfFilterDto.cs
@@ -1,6 +1,8 @@
 ﻿// This is an independent project of an individual developer. Dear PVS-Studio, please check it.
 // PVS-Studio Static Code Analyzer for C, C++, C#, and Java: http://www.viva64.com
 
+using TL;
+
 namespace TgStorage.Domain.Filters;
 
 /// <summary> Filter DTO </summary>
@@ -40,13 +42,38 @@ public sealed partial class TgEfFilterDto : TgDtoBase, ITgDto<TgEfFilterEntity, 
 		SizeType = TgEnumFileSizeType.Bytes;
 	}
 
-	#endregion
+    #endregion
 
-	#region Public and private methods
+    #region Public and private methods
 
-	public override string ToString() => $"{IsEnabled} | {FilterType} | {Name} | {Mask} | {Size} | {SizeType}";
+    public string GetStringForFilterType() => FilterType switch
+    {
+        TgEnumFilterType.SingleName => TgLocaleHelper.Instance.MenuFiltersSetSingleName,
+        TgEnumFilterType.SingleExtension => TgLocaleHelper.Instance.MenuFiltersSetSingleExtension,
+        TgEnumFilterType.MultiName => TgLocaleHelper.Instance.MenuFiltersSetMultiName,
+        TgEnumFilterType.MultiExtension => TgLocaleHelper.Instance.MenuFiltersSetMultiExtension,
+        TgEnumFilterType.MinSize => TgLocaleHelper.Instance.MenuFiltersSetMinSize,
+        TgEnumFilterType.MaxSize => TgLocaleHelper.Instance.MenuFiltersSetMaxSize,
+        _ => $"<{TgLocaleHelper.Instance.MenuFiltersError}>",
+    };
 
-	public TgEfFilterDto Copy(TgEfFilterDto dto, bool isUidCopy)
+    public override string ToConsoleString() =>
+        $"{TgDataFormatUtils.GetFormatString(Name, 20).TrimEnd(),-20} | " +
+        $"{TgDataFormatUtils.GetFormatString(Mask, 20).TrimEnd(),-20} | " +
+        $"{(IsEnabled ? "enabled" : "disabled"),-8} | " +
+        $"{GetStringForFilterType(),-20} | " +
+        $"{Size,12} | " +
+        $"{SizeType} ";
+
+    public override string ToConsoleHeaderString() =>
+        $"{TgDataFormatUtils.GetFormatString(nameof(Name), 20).TrimEnd(),-20} | " +
+        $"{TgDataFormatUtils.GetFormatString(nameof(Mask), 20).TrimEnd(),-20} | " +
+        $"{TgDataFormatUtils.GetFormatString(nameof(IsEnabled), 8).TrimEnd(),-8} | " +
+        $"{TgDataFormatUtils.GetFormatString(nameof(FilterType), 20).TrimEnd(),-20} | " +
+        $"{TgDataFormatUtils.GetFormatString(nameof(Size), 12).TrimEnd(),-12} | " +
+        $"SizeType";
+
+    public TgEfFilterDto Copy(TgEfFilterDto dto, bool isUidCopy)
 	{
 		base.Copy(dto, isUidCopy);
 		IsEnabled = dto.IsEnabled;